requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
langchain==0.3.0
langchain-openai==0.2.0
langchain-community==0.3.0
psycopg2-binary==2.9.9
pgvector==0.3.0
pypdf==4.3.0
python-dotenv==1.0.0
openai==1.50.0
pydantic-settings==2.5.0
python-multipart==0.0.9
httpx==0.27.0
PyJWT==2.9.0
